2.1.0 (#67)
* deprecate Crate#asReward. (#65)

The method is primarily used for claims internally. The method signature with the name can be confusing to folks. For folks who want to give a claim that is a Crate, use ClaimRegistrar#addClaim

* refactor: update Reward class to remove some helper functions. (#66)

These helper functions may not be valid in the future if we support immutability.
I'm opting to hard remove these functionality rather than a slow deprecation.

* chore: bump version to 2.1.0